refactor: create a fn getRetryDelay and test it
maxDelay now restrict the value of the delay after that the fuzzFactor
has been applied. This ensure that the returned delay is trully
equal or inferior the the maxDelay.
This would reduce from 3900ms to 3000ms the maxDefault based on our
current config but updating the value does not seem necessary.