fix(types): remove IIFE to correctly export Player class type
exporting a class with an IIFE change the exported type of the class
from "RxPlayer" to "typeof RxPlayer". That will provoke typechecking
errors for applications that used "RxPlayer" as a type.
This does NOT mess-up with the sideEffects config because
the global overrides, such as "patchWebkitSourceBuffer()" are executed
at the entry point for RxPlayer class (src/index and src/minimal).