Clean dist directory when calling build:all script
@Florent-Bouisset noticed that some release published in the v4 had a
`dist` directory with unnecessary v3-related artefacts inside.

This is due to the fact that `/dist` was added to our `.gitignore` file
(so it isn't tracked by git) and back-and-forth between v3 and v4 code
means that the dist directory might end up having potentially v3 stuff.

We could be adding those in the `.npmignore` to prevent them to be
published to npm, but I guess it makes more sense to just clean-up the
`dist` directory each time we redo the `build:all` script which
re-build every files and directories inside that that directory (and
which is the script we do before every release).