Fix representationListUpdate event
Turns out that because of a `===`/`!==` mix-up, the
`representationListUpdate` event (new event in v4) was never sent.

Fortunately, this event was rarely useful.
Here it is an application specifically purposed to inspect streams that
noticed a difference between anounced video `Representation`s at
`videoTrackChange`-time and what seemed to be available after some time
(due to a DRM fallback since).