Doc: Repeat some page groups name in h1
I noticed that the "h1" title of some documentation pages wasn't clear
enough on what the page was about when treated in isolation.

In particular, I think about the pages in our v3.x.x migration guide,
which had generic titles like "Constructor Options" (instead of
something like `v3 Migration guide: Constructor Options`).

Beside not giving a clear context as a title (which isn't that much
problematic), search results also were poorer as the results only echoed
the titles without providing the context in which that page appears
(its parent or next and previous page).

E.g. when searching for `throttleWhenHidden`, a result could be found in
a page that appeared through the search results as "Constructor Options"
without making it clear that it is inside the migration guide - and as
such that the `throttleWhenHidden` option does not exist anymore.

I think just making the `h1` always understandable as if it the page
could be seen in a standalone way (not in the context of the
next/previous/parent page) fixes the issue in a nice way.