tests: perform exponential backoff in integration tests
This is an implementation of an idea we had in the past to speed up
asynchronous tests.

Those often depend on some timeout, with a timeout set very high
(hundreds of milliseconds, sometimes seconds) to ensure a test pass even
on the slowest machine running up our tests.

This results in our integration tests taking a lot of time to finish
(more than 30 minutes), even more considering that those are not
parallelized, unlike our unit tests (and doing so is not straightforward
because we rely on a dependency, mocha, as a test runner and on another,
karma, as a browser launcher).

To make it less long, we wondered if we could not rely on another
strategy:
  1. Perform the check almost immediately, on fast computers it has a
     chance of very quickly passing
  2. If the check fails, wait some amount of time and re-perform the
     check.
  3. If the check fails multiple times, it surely means that the current
     device is slow, so we'll wait a longer and longer delay before
     re-performing the check.
  4. If the check still fails after some set time (or 4 seconds by
     default), fail the test.

This is basically another implementation of an "exponential backoff"
algorithm in the player (we already use this type of algorithm at
several places in the player, like for media requests for example).