On safari Mobile, don't await for data before switching to the LOADED state
Fixes #1390

We previously had a work-around specific to safari mobile when the
media element had the `playsinline` attribute where the browser would
never switch to the `LOADED` state if the media element was still in a
`paused` state.

From a recent issue (#1390), it seems to now even happen when
`playsinline` is not set.

We also add a check for whether the media element is `paused` (as if not
paused, such as when autoplay was handled, there doesn't seem to be any
issue).

It shouldn't break the previous seen case if we are to believe the
comment on top of the work-around.