api: throw and error if an application use a video element in multiple instance of the RxPlayer
The RxPlayer was not designed to run multiple instance of
RxPlayer with the same videoElement. Having multiple instance sharing
the same video element would result in various issues and unexpected
behavior.

This commit adds a check to throw an error if a videoElement is shared
between RxPlayer instances. This will provide better guidance to
application developers regarding any errors in their implementation.