Remove `isSafariMobile` mention from `main_thread` code.
When doing some hotfixes for our v4.0.0 version, we directly integrated
a browser detection check in the main code of the RxPlayer.

The normal way for the RxPlayer of handling device compatibility, is to
clearly list and document all differences in the `src/compat` directory,
not directly in the main code.

This commit fixes that by giving a name to that work-around, documentating it
and moving it in the `src/compat` directory.