[WIP] PoC for dummy media element and MSE API
This is a proof-of-concept to see if adding a mock implementation of
the video HTML5 API and MSE API subset used by the RxPlayer inside the
code of the RxPlayer seems useful and maintainable.

Motivation
----------

Initially, the idea was to implement one of the several solutions we
have in mind to pre-load a future content, this was the most far-fetched
solution, but I thought that it could still make enough sense to be
tried.
In this solution, the application would create two player instances:

  1. One linked to the true media element on the page, as usual

  2. The Other linked to our dummy media element which would preload
     and store locally (in memory? through storage APIs when available?)
     loaded contents.

     Note that here nothing will change in the RxPlayer API, it is just
     that the application will have provided to that instance our mocked
     media element - which would implement all that storage logic in its
     implementation of MSE API - instead of a regular one. This is to
     ensure a very minimal modification of the core RxPlayer code.

When the application judged that playback should begin, it can get the
preloaded data through an API of that dummy media element, call `stop`
on the RxPlayer instance with that dummy element and give the preloaded
data as a supplementary `loadVideo` option (`preloadedData`?) to the
RxPlayer instance with the real media element.

There are several complexities that are not yet handled here: most of
all we need to be careful as segments on that dummy implementation will
for now be stored in memory. Also, we will also need to provide
segment-identification metadata alongside the preloaded data so the
"real" RxPlayer is able to recognize which Adaptation/Representation has
already been loaded so that instance doesn't try to replace it.

Also a potential use case asked by applications is to preload a content
while another one is already loading. With how this solution is
currently implemented, this wouldn't be efficient, as both instances could
be loading media data at the same time without priorization mechanisms
(e.g. we would imagine that the currently-loaded content is more
important) - though I imagine this could be implemented in some way.

Other uses
----------

While doing a skeleton of it, I've realized that the MSE API outside of
segment parsing and decoding was relatively straightforward: throw when
the state or arguments is not right, send the right events etc., so I
thought that a second use case (which may well in final be our first use
case!) would be for testing.

Thanks to this implementation:

  1. we could just push fake generated content to facilitate writing
     tests (no need to generate a real content linked to the wanted
     behavior for each test)

  2. we could more easily replicate and test MSE implementation bugs
     seen on other devices and ensure they keep being tested.

Another nice use of it is that it implement a good chunk of the API used
by the RxPlayer that are only found in browser environments, which may
simplify PoCs in more restricted environments which can still run JS.