Fix Worker usage on PS4 browser
We noticed that ps4 applications were not relying on the `MULTI_THREAD`
despite being theoretically compatible with it.

Turns out that `typeof Worker` does not return `"function"` as expected
with such objects but `"object"` instead on that browser.
Thus, the RxPlayer believes that it hasn't access to the Worker API, the
necessary requirement for the `MULTI_THREAD` feature.

I suppose that this may also be true for other old Webkit-based browsers,
though it isn't for recent-ish webkitgtk-based browsers on my PC,
so for now I only enforce it for the PlayStation 4.