Do not rely on performance.now if unavailable
After #1401 (fix about the fact that `typeof Worker === "object"` on the
PlayStation 4 for some reason), I finally succeeded to run the
`MULTI_THREAD` experimental feature on a PlayStation 4 by adding one
supplementary fix.

Turns out that the `performance.now` web API, that we use to measure
time differences (Date.now() being clock-related, it's not always
appropriate for that) was not available in a Worker environment on the
PlayStation 4.

I have no idea why, as it is available in main thread, a random guess
would be that this is a remnant of mitigations against something like
spectre/meltdown but all I could find is that [they reduced its
precision](https://trac.webkit.org/changeset/226495/webkit), not removed
it from worker.

Anyway, we thankfully only relied on this API at two places, as we
voluntarily reduced those places previously to allow a synchronization
mechanism between main thread and worker.

This means that we can now easily use `Date.now()` in its place when
`performance.now()` is unavailable without much fear of the issue coming
back again in the future (our linter already ensure the API isn't used
elsewhere).

The main issue would be that timing measures would now become
time-dependant on the PlayStation 4 and thus may be affected by clock
changes, or subtle unix timestamp quirks. This should very rarely be
problematic though, and it will just happen when the experimental
`MULTI_THREAD` feature is explicitely relied on on the PlayStation 4
for now.