api: throw an error if an application use a video element in multiple instance of the RxPlayer (#1394)
* api: throw and error if an application use a video
element in multiple instance of the RxPlayer

The RxPlayer was not designed to run multiple instance of
RxPlayer with the same videoElement. Having multiple instance sharing
the same video element would result in various issues and unexpected
behavior.

This commit adds a check to throw an error if a videoElement is shared
between RxPlayer instances. This will provide better guidance to
application developers regarding any errors in their implementation.

* demo: dispose the previous RxPlayer instance before creating a new one
in the demo.

Commit 57fd4b916016eb10238b1deb956a998cebe46d2a added an error message
if the video element is reused between RxPlayer instance.
This was the case on the demo, and this commit ensure to dispose the
previous instance before creating a new one.

* add unit test and console.warn instead of throwing to prevent breaking change in API

* lint

* PR feedback

* format