fix(safari): autoplay was not working on Safari
When seeking on the current position the video element trigger the
"seeking" event, but the seeking attribute of the video element itself
is either not updated or is already reset to `false` at the moment the
event "seeking" is received.

video.onseeking = () => {
    // video.seeking attribute can be false
    console.log("on seeking!", video.seeking)
}

Due to a recent change we are waiting for the attribute "video.seeking"
to be true. Checking that the event type correspond to a seeking event
fix the issue.