[v3] Await some delay before re-attempting to push a segment following an error
This commit updates the way the RxPlayer handles a `QuotaExceededError`
error thrown by a `SourceBuffer.prototype.appendBuffer` call, to align
it better to how we do it in the v4 - which seems more appropriate.

The idea is that a browser might throw this QuotaExceededError when too
much data has been pushed to the buffer and thus the browser is not able
to accept any more segments for now.

In the v3 until now, what the RxPlayer did was trying to remove some
data from the buffer - even skipping that part if the buffer was
already empty enough - and then re-attempting to push the same segment.
If the error repeats, there is a complex chain of events that lead to
the buffering of that segment in our last v3 releases, so there's
no real issue there.

The idea of this commit is to prevent doing the re-attempt of pushing
the same segment directly, mostly for cases where the data-removal step
is skipped, by awaiting 200 milliseconds.
I also added some code ensuring that we do not retry to push the
segment if the corresponding "RepresentationStream" has been cancelled
while it was awaiting those 200 milliseconds.