When listing segments needed, stop filtering the current range
While debugging an issue, I noticed that for optimization purposes we
filtered already-loaded segments' metadata so we only considered those
of the currently-needed time range before giving it to both the
`getNeededSegments` function, which lists the segments the RxPlayer
needs to download, and the `checkForDiscontinuity` function, which
checks if a discontinuity is imminent.

This is problematic since we added the `maxVideoBufferSize` option (so a
long time ago!), as that option is also exploited in
`getNeededSegments` yet needs to know **ALL** of the buffered segments
so it can estimate the amount of video data currently retained by the
browser's video buffer - and not just those in the currently-needed
range. So it seems that the `maxVideoBufferSize`-linked estimate could
have been inexact most of the time here.